Fix client's fileGetPath() FormatException
FormatException is returned when getting path of files residing in another
resource. This happens because `CResource::GetName` is of different type on
client and server side (`const char*` vs `const SString&`).

Possibly it would be better to even out these methods but it also requires
adjustments and verification in other parts of project.